Extracted scheduler methods to separate file, adjusted check mechanism
