+++ .gitignore
@@ -1,1 +1,2 @@
 *.exe
+*-config.json
